+++ changes.md
@@ -32,6 +32,7 @@ approx 5%, for details of test see PR ([#558](https://github.com/ess-dmsc/kafka-
 - Better documentation of writer module configuration options.
 - The application will now print an error message if there is a configuration that is not used (due to e.g. a typo).
 - The error reporting and handling of writer module configurations have overall been greatly improved.
+- Added support for writing ADAr schemas.
 
 ## Command system re-write
 
@@ -40,3 +41,4 @@ approx 5%, for details of test see PR ([#558](https://github.com/ess-dmsc/kafka-
 - The code is (currently) the documentation. Instead of implementing your own code for commanding the file-writer, you can use the [file-writer-control](https://github.com/ess-dmsc/file-writer-control) Python library.
 - Furthermore, command and status topics have been merged into one command + status topic. The command line argument has thus also been changed to `--command-status-uri`.
 - There is now a job pool functionality. The job pool topic is set with the argument `--job-pool-uri`.
+